Add support to find installed vsgvr data files
To find data files when vsgvr is packaged by a
distribution, the search is extended to the
vsgvr data install location, which is relative
based on the path of the executable and determined
by the cmake buildsystem. This is required to
support a portable installation, such as those
found on Windows.

In order for the executable to find the required
data files when running from the build directory,
the location of the data files in the build
directory directory has been adjusted.